Lock screen when scanning for device
If Use CGM as heartbeat is selected, but there's been no initial scan completed yet, then lock the screen.
Loop app will scan for the device, as soon as device is connected, then the lock is disabled again.

Because if the app goes to the background, it will stop scanning (this does not apply to Dexcom transmitters, in case of Dexcom, scanning goes on even of the iOS app is in the background - but because the  client is not aware of the type of transmitters, it keeps the app in the foreground, no matter the type of CGM)