Temp basal rate when actual value is below average correction
When the actual glucose value is below the average correction range, then the basal rate can be reduced with a percentage.

Why? It's impossible to set an exactly correct basal rate. The rate I set will be either too high or too low. I prefer to estimate my basal rate too high when my values are high. Because if my basal is set too low, and my values are too high, then theoretically it's impossible to come back in range, except if I work with a ISF that is also too high, which is getting dangerous.
So I want to estimate my basal rate a little bit too high.
But once the values are in range, then I want to be able to use a lower basal rate.

With this change, it's possible to specify a temporary basal rate, being a percentage of the normal basal rate, with preconditions:

- enabled by user via UI ((set via xDrip4iOS UI)

- there's no correction and no other temp basal (for any other reason) recommended by Loop

- no other temporary basal running at the moment

- there's no basal rate schedule override active

- Loop does not take into account manual temp basal when calculating glucose effects (this is enabled/disabled via xDrip4iOS UI)
   - this is another 'personal' feature already implemented: when I set myself a temporary basal, I want to be able to tell to Loop to not take into account manual temp basal in glucose effect calculation. That other feature needs to be used, ie loop should not take into account manual temp basals in glucose effect calculation.
   - the new type of temp basal is for that reason registered as a manual temp basal.
Otherwise, Loop would simply find a negative IOB (because there's a temp basal reducing the normal basal), then would predict that future values would go up and finally this may end up in Loop starting to micro-bolus. This is not the aim off course.

   - The feature of being able to not take into account manual basals in glucose effect calculations is something already developed previously.

- latest reading should be less than 10 minutes old

If those preconditions are met, if current value is below average in correction range, and if so set temp basal at x% of actual current basal rate. X can be set by user

Disadvantages :
- it's considered as a manual temp basal, which means it shows on the Home Screen as a temp basal

Changes in several repositories :

https://github.com/JohanDegraeve/xdrip-client-swift-1/tree/mydev
- there's a new option in the UI to enable this feature, also with a percentage to apply to the basal rate (between 0 and 100, in steps of 10)

https://github.com/JohanDegraeve/Loop-LoopKit/tree/mydev2
- class DoseEnactor function enact, now supports parameter 'automatic', as explained below
- the function recommendedAutomaticDose is where the actual check is done to see if a temp basal needs to be created

https://github.com/JohanDegraeve/LoopKit-LoopKit/tree/mydev
- class TempBasalRecommendation has a new attribute named automatic, to support this functionality. Default true
- the PumpManager protocol function enactTempBasal now has a new parameter named 'automatic', which is set to false for the new type of temp basalt, true for other automatic temp basal changes

https://github.com/JohanDegraeve/rileylink_ios-LoopKit/tree/mydev
- in OmnipodPumpManager, support of automatic attribute in enactTempBasal
- removed the beeps when setting temp basal (because it's set as a manual, which normally goes with beeps when starting and stopping)

https://github.com/JohanDegraeve/OmniBLE-LoopKit/tree/mydev
- support of attribute 'automatic' in enactTempBasal