Loopshare timing fix : avoid duplicate readings
Assuming this is the cause:
Sometimes there were duplicate values visible in HealthKit.
Sometimes previous reading (of previous minute) reappears with a small timing difference off a few seconds, causing it to be readded.

Now filter is done on reading timestamp > last reading timestamp + 30 seconds.

Seems to avoid the duplicates in HK